+++ src/NAME/SqlServer/SQLServerVersionResolver.cs
@@ -75,7 +75,7 @@ namespace NAME.SqlServer
             }
 
             if (dataSource == null)
-                throw new NAMEException("The SQL Server connection string must contain a Data Source or Server value.");
+                throw new FormatException("The SQL Server connection string must contain a Data Source or Server value.");
 
             var parts = dataSource.Split(',');
             hostname = parts[0];
@@ -94,7 +94,7 @@ namespace NAME.SqlServer
                 //Read Type (should be response = 4)
                 int type = reader.ReadByte();
                 if (type != 4)
-                    throw new NAMEException($"{SupportedDependencies.SqlServer}: Server responded with wrong Type ({type}).");
+                    throw new NAMEException($"{SupportedDependencies.SqlServer}: Server responded with wrong Type ({type}).", NAMEStatusLevel.Error);
                 //Skip Status bit mask
                 reader.ReadByte();
                 //Read the full message length
@@ -110,13 +110,13 @@ namespace NAME.SqlServer
                 //Read first option token (should be Version = 0)
                 int token = reader.ReadByte();
                 if (token != 0)
-                    throw new NAMEException($"{SupportedDependencies.SqlServer}: Server responded with wrong Token ({token}).");
+                    throw new NAMEException($"{SupportedDependencies.SqlServer}: Server responded with wrong Token ({token}).", NAMEStatusLevel.Error);
                 //Read the offset
                 ushort offset = reader.ReadUInt16();
                 //Read the length (should be 6)
                 ushort optionlength = reader.ReadUInt16();
                 if (optionlength != 6)
-                    throw new NAMEException($"{SupportedDependencies.SqlServer}: Server responded with an invalid version length ({length}).");
+                    throw new NAMEException($"{SupportedDependencies.SqlServer}: Server responded with an invalid version length ({length}).", NAMEStatusLevel.Error);
                 //Skip everything until the version.
                 reader.ReadBytes(offset - 2 - 2 - 1);
                 int major = reader.ReadByte();