+++ CHANGELOG.md
@@ -13,6 +13,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 * A relevant error message is now shown when a service dependency does not have NAME installed. [(#14)](https://github.com/nosinovacao/name-sdk/issues/14)
 * Support for wildcards in the maximum version of dependencies. [(#11)](https://github.com/nosinovacao/name-sdk/issues/11)
 * Support for Elasticsearch version resolving. [(#19)](https://github.com/nosinovacao/name-sdk/issues/19)
+* A status field for each dependency in the manifest that describes its error level. [(#24)](https://github.com/nosinovacao/name-sdk/issues/24)
+* Startup healthcheck is now performed asynchronously when `ThrowOnDependenciesFail` is set to false. [(#4)](https://github.com/nosinovacao/name-sdk/issues/19)
 
 ## v1.0.0 - 2017-07-07
 * Initial open source release.